feat(delay): implement new delay utility with various asynchronous control methods and add queueMacrotask functionality
