+++ pkg/nanolib/delay/src/lib.ts
@@ -1,0 +1,4 @@
+import {getGlobalThis} from '@alwatr/global-this';
+import type {DictionaryOpt} from '@alwatr/type-helper';
+
+export const globalThis_ = getGlobalThis<DictionaryOpt<unknown>>();
